pyproject.toml
[tool.poetry]
name = "create_profile_description"
version = "0.1.0"
description = ""
authors = ["moarshy <marshath@gmail.com>"]
readme = "README.md"
packages = [
    { include = "create_profile_description", from = "." },
]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.12.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
